Add diagnostics.
Refactored `AvroSourceGenerator.cs` to improve schema handling and
diagnostic reporting. Introduced new diagnostic classes for Avro
schema-related errors and warnings. Moved `TemplateLoader` and
`NameHelper` to new files. Updated `SourceOutputModel` to use
`SchemaFieldInfo` record struct. Added new diagnostic rules in
`AnalyzerReleases.Unshipped.md`. Updated `AnalyzerReleases.Shipped.md`
with shipped analyzer releases. Introduced `AvroTemplate` class for
template rendering.
