+++ src/AvroSourceGenerator/AnalyzerReleases.Unshipped.md
@@ -1,0 +1,11 @@
+﻿; Unshipped analyzer release
+; https://github.com/dotnet/roslyn-analyzers/blob/main/src/Microsoft.CodeAnalysis.Analyzers/ReleaseTrackingAnalyzers.Help.md
+
+### New Rules
+
+Rule ID | Category | Severity | Notes
+--------|----------|----------|-------
+AVROSG0001 | Usage | Warning | MissingAvroSchemaMemberDiagnostic
+AVROSG0002 | Usage | Warning | InvalidAvroSchemaMemberDiagnostic
+AVROSG0003 | Syntax | Error | InvalidJsonSchemaDiagnostic
+AVROSG0004 | Semantic | Error | InvalidAvroSchemaDiagnostic
