+++ src/AvroSourceGenerator/Schemas/InvalidSchemaException.cs
@@ -1,0 +1,3 @@
+﻿namespace AvroSourceGenerator.Schemas;
+
+internal sealed class InvalidSchemaException(string message) : Exception(message);
