+++ src/AvroSourceGenerator/Schemas/SchemaRegistry.cs
@@ -29,14 +29,13 @@ internal sealed class SchemaRegistry(LanguageFeatures languageFeatures, string? 
             JsonValueKind.String => KnownType(schema, @namespace, nullable),
             JsonValueKind.Object => ComplexType(schema, @namespace, nullable),
             JsonValueKind.Array => Union(schema, @namespace),
-            _ => throw new InvalidOperationException($"Invalid schema {schema.GetRawText()}")
+            _ => throw new InvalidSchemaException($"Invalid schema type: {schema.GetRawText()}")
         };
     }
 
     private QualifiedName KnownType(JsonElement schema, string? @namespace, bool nullable)
     {
-        // TODO: Find a way to avoid string allocation here. Maybe go back to use internal raw value?
-        var name = schema.GetString() ?? throw new InvalidOperationException($"Invalid schema {schema.GetRawText()}");
+        var name = schema.GetString() ?? throw new InvalidSchemaException($"Invalid schema: expected a string but got {schema.GetRawText()}");
 
         return name switch
         {
@@ -49,7 +48,7 @@ internal sealed class SchemaRegistry(LanguageFeatures languageFeatures, string? 
             "bytes" => QualifiedName.Bytes(UseNullableReferenceTypes & nullable),
             "string" => QualifiedName.String(UseNullableReferenceTypes & nullable),
             _ when _schemas.TryGetValue(new QualifiedName(NameHelper.GetValid(name), @namespace), out var registeredType) => registeredType.QualifiedName,
-            _ => throw new InvalidOperationException($"Invalid schema {schema.GetRawText()}")
+            _ => throw new InvalidSchemaException($"Unknown schema type: {name}")
         };
     }
 
@@ -58,8 +57,7 @@ internal sealed class SchemaRegistry(LanguageFeatures languageFeatures, string? 
         if (schema.TryGetProperty("logicalType", out var logicalType))
             return Logical(logicalType, nullable);
 
-        // TODO: Find a way to avoid string allocation here. Maybe go back to use internal raw value?
-        var type = schema.GetProperty("type").GetString() ?? throw new InvalidOperationException($"Invalid schema {schema.GetRawText()}");
+        var type = schema.GetProperty("type").GetString() ?? throw new InvalidSchemaException($"Invalid schema: missing 'type' property in {schema.GetRawText()}");
 
         return type switch
         {
@@ -69,7 +67,7 @@ internal sealed class SchemaRegistry(LanguageFeatures languageFeatures, string? 
             "record" => Record(schema, @namespace, nullable),
             "error" => Error(schema, @namespace, nullable),
             "fixed" => Fixed(schema, @namespace, nullable),
-            _ => throw new InvalidOperationException($"Invalid schema {schema.GetRawText()}")
+            _ => throw new InvalidSchemaException($"Unknown complex schema type: {type} in {schema.GetRawText()}")
         };
     }
 
@@ -96,17 +94,16 @@ internal sealed class SchemaRegistry(LanguageFeatures languageFeatures, string? 
         if (!_schemas.TryGetValue(name, out var enumSchema))
         {
             var documentation = schema.TryGetProperty("doc", out var doc) ? doc.GetString() : null;
-            var aliases = schema.TryGetProperty("aliases", out var ali) ? ali.EnumerateArray().Select(alias => alias.GetString() ?? throw new InvalidOperationException($"Invalid schema {schema.GetRawText()}")).ToImmutableArray() : [];
-            var symbols = schema.GetProperty("symbols").EnumerateArray().Select(symbol => NameHelper.GetValid(symbol.GetString() ?? throw new InvalidOperationException($"Invalid schema {schema.GetRawText()}"))).ToImmutableArray();
+            var aliases = schema.TryGetProperty("aliases", out var ali) ? ali.EnumerateArray().Select(alias => alias.GetString() ?? throw new InvalidSchemaException($"Invalid alias in schema: {schema.GetRawText()}")).ToImmutableArray() : [];
+            var symbols = schema.GetProperty("symbols").EnumerateArray().Select(symbol => NameHelper.GetValid(symbol.GetString() ?? throw new InvalidSchemaException($"Invalid symbol in schema: {schema.GetRawText()}"))).ToImmutableArray();
             var @default = schema.TryGetProperty("default", out var def) ? def.GetString() : null;
             if (@default is not null && symbols.IndexOf(@default) is -1)
-                throw new InvalidOperationException($"Invalid schema {schema.GetRawText()}");
+                throw new InvalidSchemaException($"Default value '{@default}' not found in symbols for schema: {schema.GetRawText()}");
 
             _schemas[name] = enumSchema = new EnumSchema(schema, name, documentation, aliases, symbols, @default);
         }
 
         return nullable ? name.ToNullable() : name;
-
     }
 
     private QualifiedName Record(JsonElement schema, string? @namespace, bool nullable)
@@ -118,7 +115,7 @@ internal sealed class SchemaRegistry(LanguageFeatures languageFeatures, string? 
         if (!_schemas.TryGetValue(name, out var recordSchema))
         {
             var documentation = schema.TryGetProperty("doc", out var doc) ? doc.GetString() : null;
-            var aliases = schema.TryGetProperty("aliases", out var ali) ? ali.EnumerateArray().Select(alias => alias.GetString() ?? throw new InvalidOperationException($"Invalid schema {schema.GetRawText()}")).ToImmutableArray() : [];
+            var aliases = schema.TryGetProperty("aliases", out var ali) ? ali.EnumerateArray().Select(alias => alias.GetString() ?? throw new InvalidSchemaException($"Invalid alias in schema: {schema.GetRawText()}")).ToImmutableArray() : [];
             var fields = schema.GetProperty("fields").EnumerateArray().Select(field => Field(field, @namespace)).ToImmutableArray();
 
             _schemas[name] = recordSchema = new RecordSchema(schema, name, documentation, aliases, fields);
@@ -136,7 +133,7 @@ internal sealed class SchemaRegistry(LanguageFeatures languageFeatures, string? 
         if (!_schemas.TryGetValue(name, out var recordSchema))
         {
             var documentation = schema.TryGetProperty("doc", out var doc) ? doc.GetString() : null;
-            var aliases = schema.TryGetProperty("aliases", out var ali) ? ali.EnumerateArray().Select(alias => alias.GetString() ?? throw new InvalidOperationException($"Invalid schema {schema.GetRawText()}")).ToImmutableArray() : [];
+            var aliases = schema.TryGetProperty("aliases", out var ali) ? ali.EnumerateArray().Select(alias => alias.GetString() ?? throw new InvalidSchemaException($"Invalid alias in schema: {schema.GetRawText()}")).ToImmutableArray() : [];
             var fields = schema.GetProperty("fields").EnumerateArray().Select(field => Field(field, @namespace)).ToImmutableArray();
 
             _schemas[name] = recordSchema = new ErrorSchema(schema, name, documentation, aliases, fields);
@@ -150,7 +147,7 @@ internal sealed class SchemaRegistry(LanguageFeatures languageFeatures, string? 
         var name = NameHelper.GetLocalName(field);
         var type = Type(field.GetProperty("type"), @namespace, nullable: false);
         var documentation = field.TryGetProperty("doc", out var doc) ? doc.GetString() : null;
-        var aliases = field.TryGetProperty("aliases", out var ali) ? ali.EnumerateArray().Select(alias => alias.GetString() ?? throw new InvalidOperationException($"Invalid field {field.GetRawText()}")).ToImmutableArray() : [];
+        var aliases = field.TryGetProperty("aliases", out var ali) ? ali.EnumerateArray().Select(alias => alias.GetString() ?? throw new InvalidSchemaException($"Invalid alias in field: {field.GetRawText()}")).ToImmutableArray() : [];
         var @default = field.TryGetProperty("default", out var def) ? GetValue(type, def) : null;
         var order = field.TryGetProperty("order", out var ord) ? ord.GetInt32() : default(int?);
         return new Field(name, type, documentation, aliases, @default, order);
@@ -192,7 +189,7 @@ internal sealed class SchemaRegistry(LanguageFeatures languageFeatures, string? 
         if (!_schemas.TryGetValue(name, out var fixedSchema))
         {
             var documentation = schema.TryGetProperty("doc", out var doc) ? doc.GetString() : null;
-            var aliases = schema.TryGetProperty("aliases", out var ali) ? ali.EnumerateArray().Select(alias => alias.GetString() ?? throw new InvalidOperationException($"Invalid schema {schema.GetRawText()}")).ToImmutableArray() : [];
+            var aliases = schema.TryGetProperty("aliases", out var ali) ? ali.EnumerateArray().Select(alias => alias.GetString() ?? throw new InvalidSchemaException($"Invalid alias in schema: {schema.GetRawText()}")).ToImmutableArray() : [];
             var size = schema.GetProperty("size").GetInt32();
 
             _schemas[name] = fixedSchema = new FixedSchema(schema, name, documentation, aliases, size);
@@ -203,8 +200,7 @@ internal sealed class SchemaRegistry(LanguageFeatures languageFeatures, string? 
 
     private QualifiedName Logical(JsonElement schema, bool nullable)
     {
-        // TODO: Find a way to avoid string allocation here. Maybe go back to use internal raw value?
-        var logicalType = schema.GetString() ?? throw new InvalidOperationException($"Invalid schema {schema.GetRawText()}");
+        var logicalType = schema.GetString() ?? throw new InvalidSchemaException($"Invalid logical type in schema: {schema.GetRawText()}");
 
         return logicalType switch
         {
@@ -218,7 +214,7 @@ internal sealed class SchemaRegistry(LanguageFeatures languageFeatures, string? 
             "timestamp-micros" => QualifiedName.TimestampMicros(nullable),
             "timestamp-millis" => QualifiedName.TimestampMillis(nullable),
             "uuid" => QualifiedName.Uuid(nullable),
-            _ => throw new NotSupportedException(schema.GetRawText()),
+            _ => throw new InvalidSchemaException($"Unsupported logical type: {logicalType} in schema: {schema.GetRawText()}"),
         };
     }
 
@@ -244,8 +240,116 @@ internal sealed class SchemaRegistry(LanguageFeatures languageFeatures, string? 
             _ => QualifiedName.Object(UseNullableReferenceTypes & schema.EnumerateArray().Any(IsNull)),
         };
 
-        // TODO: Find a way to avoid string allocation here. Maybe go back to use internal raw value?
         static bool IsNull(JsonElement schema) =>
             schema.ValueKind is JsonValueKind.String && schema.GetString() == "null";
+    }
+}
+
+file static class NameHelper
+{
+    private static readonly Dictionary<string, string> s_reserved = new()
+    {
+        ["bool"] = "@bool",
+        ["byte"] = "@byte",
+        ["sbyte"] = "@sbyte",
+        ["short"] = "@short",
+        ["ushort"] = "@ushort",
+        ["int"] = "@int",
+        ["uint"] = "@uint",
+        ["long"] = "@long",
+        ["ulong"] = "@ulong",
+        ["double"] = "@double",
+        ["float"] = "@float",
+        ["decimal"] = "@decimal",
+        ["string"] = "@string",
+        ["char"] = "@char",
+        ["void"] = "@void",
+        ["object"] = "@object",
+        ["typeof"] = "@typeof",
+        ["sizeof"] = "@sizeof",
+        ["null"] = "@null",
+        ["true"] = "@true",
+        ["false"] = "@false",
+        ["if"] = "@if",
+        ["else"] = "@else",
+        ["while"] = "@while",
+        ["for"] = "@for",
+        ["foreach"] = "@foreach",
+        ["do"] = "@do",
+        ["switch"] = "@switch",
+        ["case"] = "@case",
+        ["default"] = "@default",
+        ["try"] = "@try",
+        ["catch"] = "@catch",
+        ["finally"] = "@finally",
+        ["lock"] = "@lock",
+        ["goto"] = "@goto",
+        ["break"] = "@break",
+        ["continue"] = "@continue",
+        ["return"] = "@return",
+        ["throw"] = "@throw",
+        ["public"] = "@public",
+        ["private"] = "@private",
+        ["internal"] = "@internal",
+        ["protected"] = "@protected",
+        ["static"] = "@static",
+        ["readonly"] = "@readonly",
+        ["sealed"] = "@sealed",
+        ["const"] = "@const",
+        ["fixed"] = "@fixed",
+        ["stackalloc"] = "@stackalloc",
+        ["volatile"] = "@volatile",
+        ["new"] = "@new",
+        ["override"] = "@override",
+        ["abstract"] = "@abstract",
+        ["virtual"] = "@virtual",
+        ["event"] = "@event",
+        ["extern"] = "@extern",
+        ["ref"] = "@ref",
+        ["out"] = "@out",
+        ["in"] = "@in",
+        ["is"] = "@is",
+        ["as"] = "@as",
+        ["params"] = "@params",
+        ["__arglist"] = "@__arglist",
+        ["__makeref"] = "@__makeref",
+        ["__reftype"] = "@__reftype",
+        ["__refvalue"] = "@__refvalue",
+        ["this"] = "@this",
+        ["base"] = "@base",
+        ["namespace"] = "@namespace",
+        ["using"] = "@using",
+        ["class"] = "@class",
+        ["struct"] = "@struct",
+        ["interface"] = "@interface",
+        ["enum"] = "@enum",
+        ["delegate"] = "@delegate",
+        ["checked"] = "@checked",
+        ["unchecked"] = "@unchecked",
+        ["unsafe"] = "@unsafe",
+        ["operator"] = "@operator",
+        ["explicit"] = "@explicit",
+        ["implicit"] = "@implicit",
+    };
+
+    public static string GetValid(string name) =>
+        s_reserved.TryGetValue(name, out var reserved) ? reserved : name;
+
+    public static string GetLocalName(JsonElement schema)
+    {
+        var name = schema.GetProperty("name").GetString();
+        if (string.IsNullOrWhiteSpace(name))
+        {
+            throw new InvalidSchemaException($"'name' property cannot be null or empty in the schema: {schema.GetRawText()}");
+        }
+        return GetValid(name!);
+    }
+
+    public static string? GetNamespace(JsonElement schema)
+    {
+        if (!schema.TryGetProperty("namespace", out var json))
+            return null;
+
+        return GetValid(json.GetString()!);
     }
 }
