+++ src/AvroSourceGenerator/SourceOutputModel.cs
@@ -3,10 +3,15 @@ using Microsoft.CodeAnalysis;
 
 namespace AvroSourceGenerator;
 
+internal readonly record struct SchemaFieldInfo(string SchemaJson, Location Location)
+{
+    public bool IsValid => SchemaJson is not null && Location is not null;
+}
+
 internal sealed record class SourceOutputModel(
-    string SchemaJson,
     LanguageFeatures LanguageFeatures,
     string? NamespaceOverride,
     string RecordDeclaration,
     string AccessModifier,
+    SchemaFieldInfo SchemaField,
     ImmutableArray<Diagnostic> Diagnostics);
